+++ blocklets/ai-studio/src/pages/project/import/utils.tsx
@@ -1,21 +1,24 @@
-import { NodeModel } from '@minoru/react-dnd-treeview';
+import { Template } from '../../../../api/src/store/templates';
 
-import { EntryWithMeta } from '../../../../api/src/routes/tree';
+export type TreeNode = {
+  id: string;
+  parent: string;
+  text: string;
+  data: Template;
+  type: string;
+};
 
-export type TreeNode = NodeModel<EntryWithMeta> & { type: string };
-
-const getDepTemplates = (list: TreeNode[], templateId: string, save = false) => {
-  let templates: any[] = [];
+const getDepTemplates = (list: TreeNode[], templateId: string, save = false): TreeNode[] => {
+  let templates: TreeNode[] = [];
 
   try {
-    const template = list.find((x: any) => (x.text || '').split('.')[0] === (templateId || '').split('.')[0]);
+    const template = list.find((x: TreeNode) => (x.text || '').split('.')[0] === (templateId || '').split('.')[0]);
 
     if (template) {
       if (save) {
         templates = [...templates, template];
       }
 
-      // @ts-ignore
       const nextId = template.data?.next?.id;
       if (nextId) {
         const nextTemplate = getDepTemplates(list, nextId, true);