+++ blocklets/ai-studio/CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 0.1.109 (2023-10-23)
 
-- chore: update templates id
+- chore: support discuss view
 
 ## 0.1.108 (2023-10-23)